DeckOfCards
an R6 method to simulate a Deck of Cards and a Boot of Decks. It turns out you cannot make generics for R6 classes as

"This code illustrates an important aspect of R6. Because methods are bound to individual objects, the previously created hadley object does not get this new method
"

as stated in

https://adv-r.hadley.nz/r6.html#r6-important-methods